css/css-values/calc-serialization-002.html

<!DOCTYPE html>

  <meta charset="UTF-8">

  <title>CSS Values and Units: serialization of calc() specified values: 17 arithmetical operations (complex)</title>

  <!--

  Latest update: March 31st 2019

  Mixed absolute length units in calc() specified values
  should *not* be converted into px unit when serializing
  calc() values

  More info:
  Issue 3741: [css-values-4] Should physical units in specified
  values in calc() collapse into px unit when serialized?
  https://github.com/w3c/csswg-drafts/issues/3741

  -->

  <link rel="author" title="Gérard Talbot" href="http://www.gtalbot.org/BrowserBugsSection/css21testsuite/">
  <link rel="help" href="https://www.w3.org/TR/css-values-4/#calc-serialize">

  <meta content="This test verifies how 17 arithmetical operations of mixed length units in calc() specified values are serialized. Absolute length units (cm, in, mm, pc, pt, q, px), font-relative length units (ex, em, rem), viewport-percentage length units (vh, vmax, vmin, vw) and percentage units are tested. 12 additions, 4 substractions and 1 division are tested." name="assert">

  <script src="/resources/testharness.js"></script>

  <script src="/resources/testharnessreport.js"></script>

  <div id="target"></div>

  <script>
  function startTesting()
  {

  var targetElement = document.getElementById("target");

    function verifySerialization(specified_value, serialization_expected, description)
    {

    test(function()
      {

      targetElement.style.height = specified_value;

      assert_equals(targetElement.style.height, serialization_expected);

      }, description);
    }

  /*

  "
  Sort the terms in the following order:

    The number, if present

    The percentage, if present

    The dimensions, ordered by their units ASCII case-insensitive alphabetically
  "
  https://www.w3.org/TR/css-values-4/#math-function-serialize-a-summation

  */

  /* Additions */

    verifySerialization("calc(1vh + 2px + 3%)", "calc(3% + 2px + 1vh)", "testing calc(1vh + 2px + 3%)");

    verifySerialization("calc(4px + 1vh)", "calc(4px + 1vh)", "testing calc(4px + 1vh)");

    verifySerialization("calc(5px + 6em + 1vh)", "calc(6em + 5px + 1vh)", "testing calc(5px + 6em + 1vh)");

    verifySerialization("calc(-8px + 9em + 1vh)", "calc(9em - 8px + 1vh)", "testing calc(-8px + 9em + 1vh)");

    verifySerialization("calc(1pc + 1in + 1vh + 10%)", "calc(10% + 1in + 1pc + 1vh)", "testing calc(1pc + 1in + 1vh + 10%)");

    verifySerialization("calc(25.4q + 1vh + 12%)", "calc(12% + 25.4q + 1vh)", "testing calc(25.4q + 1vh + 12%)");

    verifySerialization("calc(1em + 1.27cm + 13% + 3em)", "calc(13% + 1.27cm + 4em)", "testing calc(1em + 1.27cm + 13% + 3em)");

 /* verifySerialization(specified_value, serialization_expected, description)  */

    verifySerialization("calc(15vw + 16vmin - 17vh)", "calc(-17vh + 16vmin + 15vw)", "testing calc(15vw + 16vmin - 17vh)");

    verifySerialization("calc(8pt + calc(9rem + 10px))", "calc(8pt + 10px + 9rem)", "testing calc(8pt + calc(9rem + 10px))");
    /*
    To simplify an expression:

      1. Replace any calc() values with parentheses containing their contents.
    https://www.w3.org/TR/css-values-4/#math-function-simplify-an-expression
    (...)
    The result must be a summation of unique units
    */

    verifySerialization("calc(6pt + 5em + 4pt + 3em)", "calc(8em + 10pt)", "testing calc(6pt + 5em + 4pt + 3em)");
    /*
    To simplify an expression:

      3. Combine identical units.
    https://www.w3.org/TR/css-values-4/#math-function-simplify-an-expression
    (...)
    The result must be a summation of unique units
    */

 /* verifySerialization(specified_value, serialization_expected, description)  */

    verifySerialization("calc(4vmin + 0pt + 3pc)", "calc(3pc + 0pt + 4vmin)", "testing calc(4vmin + 0pt + 3pc)");
    /*
    Terms with a value of zero must be preserved in this summation.
    https://www.w3.org/TR/css-values-4/#math-function-simplify-an-expression
    */

    verifySerialization("calc(4vmin + 0pt)", "calc(0pt + 4vmin)", "testing calc(4vmin + 0pt)");
    /*
    Terms with a value of zero must be preserved in this summation.
    https://www.w3.org/TR/css-values-4/#math-function-simplify-an-expression
    */


  /* Substractions */


    verifySerialization("calc(1vh - 7px)", "calc(-7px + 1vh)", "testing calc(1vh - 7px)");

    verifySerialization("calc(5ex - 9ex)", "calc(-4ex)", "testing calc(5ex - 9ex)");
    /*
    To simplify an expression:

      3. Combine identical units.
    https://www.w3.org/TR/css-values-4/#math-function-simplify-an-expression

    An out-of-range value inside calc() is not syntactically invalid.
    Inside a calc() function, the resulting summation value can be
    out of range and can be serialized.
    */

    verifySerialization("calc(-80px + 25.4mm)", "calc(25.4mm - 80px)", "testing calc(-80px + 25.4mm)");
    /*
    To serialize a calc() value

      3. Otherwise, serialize the summation, prefix the result
      with "calc(" and suffix it with ")", then return it.
    https://www.w3.org/TR/css-values-4/#serialize-a-calc-value

    To serialize a summation:

      2. Serialize all the terms, then join them into a single
      string, with " + " between each term. Return the result.
    https://www.w3.org/TR/css-values-4/#math-function-serialize-a-summation
    (...)
    The result must be a summation of unique units
    */

 /* verifySerialization(specified_value, serialization_expected, description)  */

    verifySerialization("calc(1vmin - 14%)", "calc(-14% + 1vmin)", "testing calc(1vmin - 14%)");


  /* Multiplication and division */


    verifySerialization("calc(4 * 3px + 4pc / 8)", "calc(0.5pc + 12px)", "testing calc(4 * 3px + 4pc / 8)");
    /*
    To simplify an expression:

      2.Resolve all multiplications and divisions.
    https://www.w3.org/TR/css-values-4/#math-function-simplify-an-expression
    (...)
    The result must be a summation of unique units
    */

    /*
    This calc(4 * 3px + 4pc / 8) test is on purpose last. We want the
    div#target to occupy 20px and to not cause document box height
    to be unneedlessly tall.
    */

  }

  startTesting();

  </script>
