third_party/blink/web_tests/external/wpt/css/css-values/calc-serialization-002-expected.txt

This is a testharness.js-based test.
[FAIL] testing calc(1vh + 2px + 3%)
  assert_equals: expected "calc(3% + 2px + 1vh)" but got "calc((3% + 2px) + 1vh)"
[FAIL] testing calc(5px + 6em + 1vh)
  assert_equals: expected "calc(6em + 5px + 1vh)" but got "calc((6em + 5px) + 1vh)"
[FAIL] testing calc(-8px + 9em + 1vh)
  assert_equals: expected "calc(9em - 8px + 1vh)" but got "calc((9em - 8px) + 1vh)"
[FAIL] testing calc(1pc + 1in + 1vh + 10%)
  assert_equals: expected "calc(10% + 112px + 1vh)" but got "calc((10% + 112px) + 1vh)"
[FAIL] testing calc(25.4q + 1vh + 12%)
  assert_equals: expected "calc(12% + 24px + 1vh)" but got "calc((12% + 24px) + 1vh)"
[FAIL] testing calc(1em + 1.27cm + 13% + 3em)
  assert_equals: expected "calc(13% + 4em + 48px)" but got "calc((13% + 4em) + 48px)"
[FAIL] testing calc(15vw + 16vmin - 17vh)
  assert_equals: expected "calc(-17vh + 16vmin + 15vw)" but got "calc((-17vh + 16vmin) + 15vw)"
Harness: the test ran to completion.
